backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
ghana-nlp
pydantic
python-multipart
